+++ websocket.h
@@ -2,6 +2,7 @@
 #define WEBSOCKET_H
 
 #include <libwebsockets.h>
+#include <stdint.h>
 
 struct per_session_data {
     struct lws *wsi;
@@ -12,7 +13,7 @@ extern struct lws *client_wsi;
 
 int callback_websocket(struct lws *wsi, enum lws_callback_reasons reason, void *user, void *in, size_t len);
 void send_message_to_client(const char *key);
-int init_websocket_server(int port);
+int init_websocket_server(uint16_t port);
 void destroy_websocket_server();
 
 extern struct lws_protocols protocols[];
